Implement bitonic_join builtin
This commit adds a new compiler built-in function `bitonic_join`.
Contrary to the `for-join` loop, this function returns the elements
of the join as an array. It is generic over its input arrays element
types and sizes. If the input arrays contain tuples, the first element
of the tuples must be same type and will be used to join the arrays on.
If the element type is not a tuple, the join is equivalent to a set
intersection. The signature is either:

```
bitonic_join([A; SIZE_0], [A; SIZE_1]) -> [(bool, A); const { SIZE_0 + SIZE_1 - 1}]
```

when A is not a tuple type. Or:
```
bitonic_join([A; SIZE_0], [B; SIZE_1]) -> [(bool, A, B); const { SIZE_0 + SIZE_1 - 1}]
```
when A and B are tuples and the types of A.0 and B.0 are equal.